backend/requirements.txt
fastapi[standard]>=0.114.0,<0.116.1
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
supabase>=2.17.0
python-dotenv>=1.1.0
pydantic>=2.11.7
PyJWT>=2.10.1
cryptography>=45.0.6
requests>=2.32.4
loguru>=0.7.3
httpx>=0.28.1
pytest>=8.4.1